Add files via upload
Final project for my 212 course in Java. The goal of this project was to implement a linked list to sort a list of dates that were scanned from a separate .txt file. When ran, it should open a GUI where the dates are displayed in two sections; unsorted and sorted.